Moving the PatientExtractor class into the mapping package.
