+++ boggs/src/test/java/com/hartwig/healthchecks/boggs/io/PatientExtractorTest.java
@@ -1,6 +1,7 @@
 package com.hartwig.healthchecks.boggs.io;
 
 import com.google.common.io.Resources;
+import com.hartwig.healthchecks.boggs.healthcheck.mapping.PatientExtractor;
 import com.hartwig.healthchecks.boggs.model.PatientData;
 import com.hartwig.healthchecks.boggs.flagstatreader.FlagStatData;
 import com.hartwig.healthchecks.boggs.flagstatreader.FlagStatParser;
